Address issue with blog images (#69)
The correct URL (or at least a redirect to it) for full-size blog post images is actually under "original_url" rather than "url" in the json.  Unfortunately, the prior version of process_content_type won't work on it since it doesn't actually link directly to the image (it won't detect the mimetype, and thus won't assign an extension).  I modified process_content_type slightly so it should now work as intended in all cases.